+++ src/components/entities/PlaceCard/index.ts
