+++ src/components/routes/PrivateRoute.tsx
@@ -1,5 +1,7 @@
 import { Navigate } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
+
+import { AppRoute } from '../shared/config/route';
+import { AuthorizationStatus } from '../shared/config/auth-status';
 
 type PrivateRouteProps = {
   children: JSX.Element;
