+++ src/components/shared/config/auth-status.ts
@@ -1,0 +1,5 @@
+export enum AuthorizationStatus {
+  Auth = 'AUTH',
+  NoAuth = 'NO_AUTH',
+  Unknown = 'UNKNOWN'
+}
