+++ src/components/shared/config/route.ts
@@ -1,0 +1,6 @@
+export enum AppRoute {
+  Root = '/',
+  Login = '/login',
+  Favorites = '/favorites',
+  Offer = '/offer/:offerId',
+}
