+++ src/components/shared/types/Orientation.type.ts
