--- src/data/citiesData.json
@@ -1,278 +1,0 @@
-[
-  {
-    "id": "14efc878-0d8a-4270-9667-bf5241632223",
-    "title": "The Joshua Tree House",
-    "type": "hotel",
-    "price": 421,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/8.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.36554,
-      "longitude": 4.911976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": false,
-    "rating": 5
-  },
-  {
-    "id": "d08a53c2-33fc-4c09-8bb7-cdc6bef81a04",
-    "title": "The house among olive ",
-    "type": "room",
-    "price": 123,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/18.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.385540000000006,
-      "longitude": 4.902976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": true,
-    "rating": 4.2
-  },
-  {
-    "id": "59f20fe1-4df0-404e-b9e5-9bf74a62f489",
-    "title": "Loft Studio in the Central Area",
-    "type": "hotel",
-    "price": 337,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/13.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.397540000000006,
-      "longitude": 4.9099759999999995,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": false,
-    "rating": 1.6
-  },
-  {
-    "id": "7007da91-bc5c-4ffc-ba45-69e7deb300c0",
-    "title": "House in countryside",
-    "type": "room",
-    "price": 280,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/12.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.37454,
-      "longitude": 4.881976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": false,
-    "rating": 4.6
-  },
-  {
-    "id": "10a48892-0142-4656-b15c-9c74b28ef689",
-    "title": "The Joshua Tree House",
-    "type": "house",
-    "price": 323,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/19.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.367540000000005,
-      "longitude": 4.883976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": true,
-    "rating": 2.2
-  },
-  {
-    "id": "4b75035c-2432-4c67-b83c-eeae702a2221",
-    "title": "Penthouse, 4-5 rooms + 5 balconies",
-    "type": "room",
-    "price": 211,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/18.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.361540000000005,
-      "longitude": 4.883976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": false,
-    "rating": 2.2
-  },
-  {
-    "id": "b051300f-e3ab-48db-ae2b-93cf6a571de3",
-    "title": "Wood and stone place",
-    "type": "apartment",
-    "price": 392,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/11.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.35754,
-      "longitude": 4.9179759999999995,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": false,
-    "rating": 2.5
-  },
-  {
-    "id": "be81e619-a68b-48fe-9eeb-dd3a0c05ef71",
-    "title": "Penthouse, 4-5 rooms + 5 balconies",
-    "type": "room",
-    "price": 206,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/20.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.37854,
-      "longitude": 4.894976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": true,
-    "rating": 3.8
-  },
-  {
-    "id": "a4fac7b3-dfc1-4d89-a55c-9de6a9029ced",
-    "title": "Loft Studio in the Central Area",
-    "type": "hotel",
-    "price": 339,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/9.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.36854,
-      "longitude": 4.887976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": false,
-    "rating": 2.3
-  },
-  {
-    "id": "db29f201-da67-4f22-9dcf-75a86ebb5629",
-    "title": "Wood and stone place",
-    "type": "hotel",
-    "price": 487,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/3.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.370540000000005,
-      "longitude": 4.9099759999999995,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": true,
-    "rating": 3.8
-  },
-  {
-    "id": "0c537de4-95c1-4479-b080-5bfeeb7872fe",
-    "title": "Nice, cozy, warm big bed apartment",
-    "type": "house",
-    "price": 398,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/10.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.389540000000004,
-      "longitude": 4.883976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": false,
-    "rating": 4.4
-  },
-  {
-    "id": "d3ebaa52-7e36-4926-9284-e8afb1b44099",
-    "title": "Waterfront with extraordinary view",
-    "type": "hotel",
-    "price": 261,
-    "previewImage": "https://14.design.htmlacademy.pro/static/hotel/11.jpg",
-    "city": {
-      "name": "Amsterdam",
-      "location": {
-        "latitude": 52.37454,
-        "longitude": 4.897976,
-        "zoom": 13
-      }
-    },
-    "location": {
-      "latitude": 52.35054,
-      "longitude": 4.908976,
-      "zoom": 16
-    },
-    "isFavorite": false,
-    "isPremium": false,
-    "rating": 4.8
-  }
-]
